Consider splitting compiler.ts into phase modules [L] (#355)
